Add basic and working master sync engine
