+++ IceCream/Classes/MasterSyncEngine.swift
@@ -1,0 +1,406 @@
+//  Created by David Collado on 1/5/18.
+//  Copyright © 2018 蔡越. All rights reserved.
+//
+
+import Foundation
+import RealmSwift
+import CloudKit
+
+public enum Notifications: String, NotificationName {
+    case cloudKitDataDidChangeRemotely
+}
+
+public enum IceCreamKey: String {
+    /// Tokens
+    case databaseChangesTokenKey
+    case zoneChangesTokenKey
+    
+    /// Flags
+    case subscriptionIsLocallyCachedKey
+    case hasCustomZoneCreatedKey
+    
+    public var value: String {
+        return "icecream.keys." + rawValue
+    }
+}
+
+/// Dangerous part:
+/// In most cases, you should not change the string value cause it is related to user settings.
+/// e.g.: the cloudKitSubscriptionID, if you don't want to use "private_changes" and use another string. You should remove the old subsription first.
+/// Or your user will not save the same subscription again. So you got trouble.
+/// The right way is remove old subscription first and then save new subscription.
+public struct IceCreamConstant {
+    public static let cloudKitSubscriptionID = "private_changes"
+}
+
+protocol SyncEnginable: class {
+    func registerLocalDatabase()
+    func cleanUp()
+    var customZoneID: CKRecordZoneID { get }
+    var zoneChangesToken: CKServerChangeToken? { get set }
+    var recordType: String { get }
+    func add(record: CKRecord)
+    func delete(recordID: CKRecordID)
+    var isCustomZoneCreated: Bool { get }
+    var syncEngineMasterDelegate: SyncEngineMasterDelegate? { get set }
+}
+
+public final class MasterSyncEngine: SyncEngineMasterDelegate {
+    
+    /// Notifications are delivered as long as a reference is held to the returned notification token. You should keep a strong reference to this token on the class registering for updates, as notifications are automatically unregistered when the notification token is deallocated.
+    /// For more, reference is here: https://realm.io/docs/swift/latest/#notifications
+    private var notificationToken: NotificationToken?
+    
+    /// Indicates the private database in default container
+    private let privateDatabase = CKContainer.default().privateCloudDatabase
+    
+    private let errorHandler = ErrorHandler()
+    
+    private let syncEngines: [SyncEnginable]
+    
+    /// We recommand process the initialization when app launches
+    public init(syncEngines: [AnyObject]) {
+        // Added this so SyncEnginable could be internal and not private
+        self.syncEngines = syncEngines.compactMap { $0 as? SyncEnginable }
+        for syncEngine in self.syncEngines {
+            syncEngine.syncEngineMasterDelegate = self
+        }
+        assert(self.syncEngines.count == syncEngines.count, "One of the proviced sync engines does not match required protocol SyncEnginable")
+        /// Check iCloud status so that we can go on
+        CKContainer.default().accountStatus { [weak self] (status, error) in
+            guard let `self` = self else { return }
+            if status == CKAccountStatus.available {
+                
+                /// 1. Fetch changes in the Cloud
+                /// Apple suggests that we should fetch changes in database, *especially* the very first launch.
+                /// But actually, there **might** be some rare unknown and weird reason that the data is not synced between muilty devices.
+                /// So I suggests fetch changes in database everytime app launches.
+                `self`.fetchChangesInDatabase({
+                    print("First sync done!")
+                })
+                
+                `self`.resumeLongLivedOperationIfPossible()
+                
+                `self`.createCustomZones(nil)
+                
+                `self`.startObservingRemoteChanges()
+                
+                /// 2. Register to local database
+                DispatchQueue.main.async {
+                    for syncEngine in `self`.syncEngines {
+                        syncEngine.registerLocalDatabase()
+                    }
+                }
+                
+                NotificationCenter.default.addObserver(self, selector: #selector(self.cleanUp), name: .UIApplicationWillTerminate, object: nil)
+                
+                if `self`.subscriptionIsLocallyCached { return }
+                `self`.createDatabaseSubscription()
+                
+            } else {
+                /// Handle when user account is not available
+                print("Easy, my boy. You haven't logged into iCloud account on your device/simulator yet.")
+            }
+        }
+    }
+    
+    // Manually sync data with CloudKit
+    public func sync() {
+        self.fetchChangesInDatabase()
+    }
+    
+    /// Create new custom zones
+    /// You can(but you shouldn't) invoke this method more times, but the CloudKit is smart and will handle that for you
+    private func createCustomZones(_ completion: ((Error?) -> ())? = nil) {
+        let zones = syncEngines.filter { !$0.isCustomZoneCreated }.map { CKRecordZone(zoneID: $0.customZoneID) }
+        let modifyOp = CKModifyRecordZonesOperation(recordZonesToSave: zones, recordZoneIDsToDelete: nil)
+        modifyOp.modifyRecordZonesCompletionBlock = { [weak self](_, _, error) in
+            guard let `self` = self else { return }
+            switch `self`.errorHandler.resultType(with: error) {
+            case .success:
+                DispatchQueue.main.async {
+                    completion?(nil)
+                }
+            case .retry(let timeToWait, _):
+                `self`.errorHandler.retryOperationIfPossible(retryAfter: timeToWait, block: {
+                    `self`.createCustomZones(completion)
+                })
+            default:
+                return
+            }
+        }
+        
+        privateDatabase.add(modifyOp)
+    }
+    
+    @objc func cleanUp() {
+        for syncEngine in syncEngines {
+            syncEngine.cleanUp()
+        }
+    }
+    
+    /// Sync local data to CloudKit
+    /// For more about the savePolicy: https://developer.apple.com/documentation/cloudkit/ckrecordsavepolicy
+    func syncRecordsToCloudKit(recordsToStore: [CKRecord], recordIDsToDelete: [CKRecordID], completion: ((Error?) -> ())? = nil) {
+        let modifyOpe = CKModifyRecordsOperation(recordsToSave: recordsToStore, recordIDsToDelete: recordIDsToDelete)
+        
+        if #available(iOS 11.0, *) {
+            let config = CKOperationConfiguration()
+            config.isLongLived = true
+            modifyOpe.configuration = config
+        } else {
+            // Fallback on earlier versions
+            modifyOpe.isLongLived = true
+        }
+        
+        // We use .changedKeys savePolicy to do unlocked changes here cause my app is contentious and off-line first
+        // Apple suggests using .ifServerRecordUnchanged save policy
+        // For more, see Advanced CloudKit(https://developer.apple.com/videos/play/wwdc2014/231/)
+        modifyOpe.savePolicy = .changedKeys
+        
+        // To avoid CKError.partialFailure, make the operation atomic (if one record fails to get modified, they all fail)
+        // If you want to handle partial failures, set .isAtomic to false and implement CKOperationResultType .fail(reason: .partialFailure) where appropriate
+        modifyOpe.isAtomic = true
+        
+        modifyOpe.modifyRecordsCompletionBlock = {
+            [weak self]
+            (_, _, error) in
+            
+            guard let `self` = self else { return }
+            
+            switch `self`.errorHandler.resultType(with: error) {
+            case .success:
+                DispatchQueue.main.async {
+                    completion?(nil)
+                    
+                    /// Cause we will get a error when there is very empty in the cloudKit dashboard
+                    /// which often happen when users first launch your app.
+                    /// So, we put the subscription process here when we sure there is a record type in CloudKit.
+                    if `self`.subscriptionIsLocallyCached { return }
+                    `self`.createDatabaseSubscription()
+                }
+            case .retry(let timeToWait, _):
+                `self`.errorHandler.retryOperationIfPossible(retryAfter: timeToWait) {
+                    `self`.syncRecordsToCloudKit(recordsToStore: recordsToStore, recordIDsToDelete: recordIDsToDelete, completion: completion)
+                }
+            case .chunk:
+                /// CloudKit says maximum number of items in a single request is 400.
+                /// So I think 300 should be a fine by them.
+                let chunkedRecords = recordsToStore.chunkItUp(by: 300)
+                for chunk in chunkedRecords {
+                    `self`.syncRecordsToCloudKit(recordsToStore: chunk, recordIDsToDelete: recordIDsToDelete, completion: completion)
+                }
+            default:
+                return
+            }
+        }
+        
+        privateDatabase.add(modifyOpe)
+    }
+}
+
+/// Chat to the CloudKit API directly
+extension MasterSyncEngine {
+    
+    /// The changes token, for more please reference to https://developer.apple.com/videos/play/wwdc2016/231/
+    var databaseChangeToken: CKServerChangeToken? {
+        get {
+            /// For the very first time when launching, the token will be nil and the server will be giving everything on the Cloud to client
+            /// In other situation just get the unarchive the data object
+            guard let tokenData = UserDefaults.standard.object(forKey: IceCreamKey.databaseChangesTokenKey.value) as? Data else { return nil }
+            return NSKeyedUnarchiver.unarchiveObject(with: tokenData) as? CKServerChangeToken
+        }
+        set {
+            guard let n = newValue else {
+                UserDefaults.standard.removeObject(forKey: IceCreamKey.databaseChangesTokenKey.value)
+                return
+            }
+            let data = NSKeyedArchiver.archivedData(withRootObject: n)
+            UserDefaults.standard.set(data, forKey: IceCreamKey.databaseChangesTokenKey.value)
+        }
+    }
+    
+    /// Cuz we only need to do subscription once succeed
+    var subscriptionIsLocallyCached: Bool {
+        get {
+            guard let flag = UserDefaults.standard.object(forKey: IceCreamKey.subscriptionIsLocallyCachedKey.value) as? Bool  else { return false }
+            return flag
+        }
+        set {
+            UserDefaults.standard.set(newValue, forKey: IceCreamKey.subscriptionIsLocallyCachedKey.value)
+        }
+    }
+    
+    /// Only update the changeToken when fetch process completes
+    private func fetchChangesInDatabase(_ callback: (() -> Void)? = nil) {
+        
+        let changesOperation = CKFetchDatabaseChangesOperation(previousServerChangeToken: databaseChangeToken)
+        
+        /// For more, see the source code, it has the detailed explanation
+        changesOperation.fetchAllChanges = true
+        
+        changesOperation.changeTokenUpdatedBlock = { [weak self] newToken in
+            guard let `self` = self else { return }
+            self.databaseChangeToken = newToken
+        }
+        
+        /// Cuz we only have one custom zone, so we don't need to store the CKRecordZoneID temporarily
+        /*
+         changesOperation.recordZoneWithIDChangedBlock = { [weak self] zoneID in
+         guard let `self` = self else { return }
+         `self`.changedRecordZoneID = zoneID
+         }
+         */
+        changesOperation.fetchDatabaseChangesCompletionBlock = {
+            [weak self]
+            newToken, _, error in
+            guard let `self` = self else { return }
+            switch `self`.errorHandler.resultType(with: error) {
+            case .success:
+                `self`.databaseChangeToken = newToken
+                // Fetch the changes in zone level
+                `self`.fetchChangesInZones(callback)
+            case .retry(let timeToWait, _):
+                `self`.errorHandler.retryOperationIfPossible(retryAfter: timeToWait, block: {
+                    `self`.fetchChangesInDatabase(callback)
+                })
+            case .recoverableError(let reason, _):
+                switch reason {
+                case .changeTokenExpired:
+                    /// The previousServerChangeToken value is too old and the client must re-sync from scratch
+                    `self`.databaseChangeToken = nil
+                    `self`.fetchChangesInDatabase(callback)
+                default:
+                    return
+                }
+            default:
+                return
+            }
+        }
+        privateDatabase.add(changesOperation)
+    }
+    
+    private var zoneIds: [CKRecordZoneID] {
+        return syncEngines.map { $0.customZoneID }
+    }
+    
+    private var zoneIdOptions: [CKRecordZoneID: CKFetchRecordZoneChangesOptions] {
+        return syncEngines.reduce([CKRecordZoneID: CKFetchRecordZoneChangesOptions]()) { (dict, syncEngine) -> [CKRecordZoneID: CKFetchRecordZoneChangesOptions] in
+            var dict = dict
+            let zoneChangesOptions = CKFetchRecordZoneChangesOptions()
+            zoneChangesOptions.previousServerChangeToken = syncEngine.zoneChangesToken
+            dict[syncEngine.customZoneID] = zoneChangesOptions
+            return dict
+        }
+    }
+    
+    private func fetchChangesInZones(_ callback: (() -> Void)? = nil) {
+        let changesOp = CKFetchRecordZoneChangesOperation(recordZoneIDs: zoneIds, optionsByRecordZoneID: zoneIdOptions)
+        changesOp.fetchAllChanges = true
+        
+        changesOp.recordZoneChangeTokensUpdatedBlock = { [weak self] zoneId, token, _ in
+            guard let `self` = self else { return }
+            guard let recordSyncEngine = `self`.syncEngines.first(where: { $0.customZoneID == zoneId }) else { return }
+            recordSyncEngine.zoneChangesToken = token
+        }
+        
+        changesOp.recordChangedBlock = { [weak self] record in
+            /// The Cloud will return the modified record since the last zoneChangesToken, we need to do local cache here.
+            /// Handle the record:
+            guard let `self` = self else { return }
+            guard let recordSyncEngine = `self`.syncEngines.first(where: { $0.recordType == record.recordType }) else { return }
+            recordSyncEngine.add(record: record)
+        }
+        
+        changesOp.recordWithIDWasDeletedBlock = { [weak self] recordId, _ in
+            guard let `self` = self else { return }
+            guard let recordSyncEngine = `self`.syncEngines.first(where: { $0.customZoneID == recordId.zoneID }) else { return }
+            recordSyncEngine.delete(recordID: recordId)
+        }
+        
+        changesOp.recordZoneFetchCompletionBlock = { [weak self](zoneId ,token, _, _, error) in
+            guard let `self` = self else { return }
+            switch `self`.errorHandler.resultType(with: error) {
+            case .success:
+                guard let recordSyncEngine = `self`.syncEngines.first(where: { $0.customZoneID == zoneId }) else { return }
+                recordSyncEngine.zoneChangesToken = token
+                callback?()
+                print("Sync successfully!")
+            case .retry(let timeToWait, _):
+                `self`.errorHandler.retryOperationIfPossible(retryAfter: timeToWait, block: {
+                    `self`.fetchChangesInZones(callback)
+                })
+            case .recoverableError(let reason, _):
+                switch reason {
+                case .changeTokenExpired:
+                    /// The previousServerChangeToken value is too old and the client must re-sync from scratch
+                    guard let recordSyncEngine = `self`.syncEngines.first(where: { $0.customZoneID == zoneId }) else { return }
+                    recordSyncEngine.zoneChangesToken = nil
+                    `self`.fetchChangesInZones(callback)
+                default:
+                    return
+                }
+            default:
+                return
+            }
+        }
+        
+        privateDatabase.add(changesOp)
+    }
+    
+    fileprivate func createDatabaseSubscription() {
+        // The direct below is the subscribe way that Apple suggests in CloudKit Best Practices(https://developer.apple.com/videos/play/wwdc2016/231/) , but it doesn't work here in my place.
+        
+        let subscription = CKDatabaseSubscription(subscriptionID: IceCreamConstant.cloudKitSubscriptionID)
+        
+        let notificationInfo = CKNotificationInfo()
+        notificationInfo.shouldSendContentAvailable = true // Silent Push
+        
+        subscription.notificationInfo = notificationInfo
+        
+        let createOp = CKModifySubscriptionsOperation(subscriptionsToSave: [subscription], subscriptionIDsToDelete: [])
+        createOp.modifySubscriptionsCompletionBlock = { _, _, error in
+            guard error == nil else { return }
+            self.subscriptionIsLocallyCached = true
+        }
+        createOp.qualityOfService = .utility
+        privateDatabase.add(createOp)
+    }
+    
+    fileprivate func startObservingRemoteChanges() {
+        NotificationCenter.default.addObserver(forName: Notifications.cloudKitDataDidChangeRemotely.name, object: nil, queue: OperationQueue.main, using: { [weak self](_) in
+            guard let `self` = self else { return }
+            `self`.fetchChangesInDatabase()
+        })
+    }
+}
+
+/// Long-lived Manipulation
+extension MasterSyncEngine {
+    /// The CloudKit Best Practice is out of date, now use this:
+    /// https://developer.apple.com/documentation/cloudkit/ckoperation
+    /// Which problem does this func solve? E.g.:
+    /// 1.(Offline) You make a local change, involve a operation
+    /// 2. App exits or ejected by user
+    /// 3. Back to app again
+    /// The operation resumes! All works like a magic!
+    fileprivate func resumeLongLivedOperationIfPossible () {
+        CKContainer.default().fetchAllLongLivedOperationIDs { ( opeIDs, error) in
+            guard error == nil else { return }
+            guard let ids = opeIDs else { return }
+            for id in ids {
+                CKContainer.default().fetchLongLivedOperation(withID: id, completionHandler: { (ope, error) in
+                    guard error == nil else { return }
+                    if let modifyOp = ope as? CKModifyRecordsOperation {
+                        modifyOp.modifyRecordsCompletionBlock = { (_,_,_) in
+                            print("Resume modify records success!")
+                        }
+                        CKContainer.default().add(modifyOp)
+                    }
+                })
+            }
+        }
+    }
+}
+
+
